OrderProcessing/requirements.txt
kafka-python==2.0.2
cassandra-driver==3.28.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2 